fix(kustomize): move logic into correct location
moves logic into classes where they make sense. for example, the code
for downloading a temp directory structure only makes sense in the
context of kustomize.
